requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
sentence-transformers>=2.5.1
numpy>=1.24.0
PyPDF2>=3.0.0
apscheduler>=3.10.4
openai>=1.6.0
jinja2>=3.1.2
